oops, fix the test
