+++ tests/appsec/iast/sink/test_hardcoded_secrets.py
@@ -4,10 +4,30 @@
 
 from utils import interfaces, weblog, features, context, missing_feature
 
-# Test_HardcodedSecrets doesn't inherit from BaseSinkTest
+# Test_HardcodedSecrets and Test_HardcodedSecretsExtended don't inherit from BaseSinkTest
 # Hardcode secrets detection implementation change a lot between different languages
 # current BaseSinkTest implementation doesn't work for all languages
 # as the vulnerability is not always set in the current request span.
+
+def get_hardcoded_secret_vulnerabilities():
+    spans = [s for _, s in interfaces.library.get_root_spans()]
+    assert spans, "No spans found"
+    spans_meta = [span.get("meta") for span in spans]
+    assert spans_meta, "No spans meta found"
+    iast_events = [meta.get("_dd.iast.json") for meta in spans_meta if meta.get("_dd.iast.json")]
+    assert iast_events, "No iast events found"
+    vulnerabilities = [event.get("vulnerabilities") for event in iast_events if event.get("vulnerabilities")]
+    assert vulnerabilities, "No vulnerabilities found"
+    vulnerabilities = sum(vulnerabilities, [])  # set all the vulnerabilities in a single list
+    hardcoded_secrets = [vuln for vuln in vulnerabilities if vuln.get("type") == "HARDCODED_SECRET"]
+    assert hardcoded_secrets, "No hardcoded secrets found"
+    return hardcoded_secrets
+
+def get_expectation(d):
+    expected = d.get(context.library.library)
+    if isinstance(expected, dict):
+        expected = expected.get(context.weblog_variant)
+    return expected
 
 
 @features.iast_sink_hardcoded_secrets
@@ -24,28 +44,28 @@ class Test_HardcodedSecrets:
 
     def test_hardcoded_secrets_exec(self):
         assert self.r_hardcoded_secrets_exec.status_code == 200
-        hardcode_secrets = self.get_hardcoded_secret_vulnerabilities()
+        hardcode_secrets = get_hardcoded_secret_vulnerabilities()
         hardcode_secrets = [v for v in hardcode_secrets if v["evidence"]["value"] == "aws-access-token"]
         assert len(hardcode_secrets) == 1
         vuln = hardcode_secrets[0]
-        assert vuln["location"]["path"] == self._get_expectation(self.location_map)
+        assert vuln["location"]["path"] == get_expectation(self.location_map)
 
-    def get_hardcoded_secret_vulnerabilities(self):
-        spans = [s for _, s in interfaces.library.get_root_spans()]
-        assert spans, "No spans found"
-        spans_meta = [span.get("meta") for span in spans]
-        assert spans_meta, "No spans meta found"
-        iast_events = [meta.get("_dd.iast.json") for meta in spans_meta if meta.get("_dd.iast.json")]
-        assert iast_events, "No iast events found"
-        vulnerabilities = [event.get("vulnerabilities") for event in iast_events if event.get("vulnerabilities")]
-        assert vulnerabilities, "No vulnerabilities found"
-        vulnerabilities = sum(vulnerabilities, [])  # set all the vulnerabilities in a single list
-        hardcoded_secrets = [vuln for vuln in vulnerabilities if vuln.get("type") == "HARDCODED_SECRET"]
-        assert hardcoded_secrets, "No hardcoded secrets found"
-        return hardcoded_secrets
+@features.iast_sink_hardcoded_secrets
+class Test_HardcodedSecretsExtended:
+    """Test Hardcoded secrets extended detection."""
 
-    def _get_expectation(self, d):
-        expected = d.get(context.library.library)
-        if isinstance(expected, dict):
-            expected = expected.get(context.weblog_variant)
-        return expected
+    location_map = {
+        "nodejs": {"express4": "iast/index.js", "express4-typescript": "iast.ts", "uds-express4": "iast/index.js"},
+    }
+
+    def setup_hardcoded_secrets_extended_exec(self):
+        self.r_hardcoded_secrets_exec = weblog.get("/iast/hardcoded_secrets_extended/test_insecure")
+
+    def test_hardcoded_secrets_extended_exec(self):
+        assert self.r_hardcoded_secrets_exec.status_code == 200
+        hardcoded_secrets = get_hardcoded_secret_vulnerabilities()
+        hardcoded_secrets = [v for v in hardcoded_secrets if v["evidence"]["value"] == "datadog-access-token"]
+        assert len(hardcoded_secrets) == 1
+        vuln = hardcoded_secrets[0]
+        assert vuln["location"]["path"] == get_expectation(self.location_map)
+
